Add cell wrapper
