+++ client/src/components/Markdown/Editor/CellEditor.vue
@@ -5,41 +5,7 @@
                 <CellButtonAdd :cell-index="cellIndex" @click="onAdd" />
             </div>
             <hr class="solid m-0" />
-            <div class="d-flex my-1 mx-2">
-                <div class="d-flex flex-column justify-content-between">
-                    <div class="cell-guide small text-primary">
-                        {{ cell.name }}
-                    </div>
-                    <div
-                        v-b-tooltip.right
-                        role="button"
-                        tabindex="0"
-                        class="d-inline text-muted cursor-pointer mt-4 align-self-end"
-                        title="Expand to Edit">
-                        <FontAwesomeIcon class="text-primary" :icon="faAngleDoubleDown" />
-                        <FontAwesomeIcon class="text-primary" :icon="faAngleDoubleUp" />
-                    </div>
-                </div>
-                <div class="ml-2 w-100">
-                    <MarkdownDefault v-if="cell.name === 'markdown'" :content="cell.content" />
-                    <MarkdownVega v-else-if="cell.name === 'vega'" :content="cell.content" />
-                    <MarkdownVitessce v-else-if="cell.name === 'vitessce'" :content="cell.content" />
-                </div>
-            </div>
-            <div class="d-flex my-1 mx-3">
-                <div class="cell-guide d-flex flex-column justify-content-between">
-                    <b-button
-                        v-b-tooltip.right
-                        class="border-0 p-0 m-0"
-                        title="Delete"
-                        variant="outline-primary">
-                        <FontAwesomeIcon :icon="faTrash" />
-                    </b-button>
-                </div>
-                <div class="ml-2 w-100">
-                    <CellCode :value="cell.content" :mode="getMode(cell.name)" @change="onChange(cellIndex, $event)" />
-                </div>
-            </div>
+            <CellWrapper :name="cell.name" :content="cell.content" @change="onChange(cellIndex, $event)" />
             <hr class="solid m-0" />
         </div>
         <CellButtonAdd :cell-index="cells.length" @click="onAdd" />
@@ -51,14 +17,7 @@ import { ref } from "vue";
 
 import { parseMarkdown } from "@/components/Markdown/parse";
 
-import { faTrash, faAngleDoubleDown, faAngleDoubleUp } from "@fortawesome/free-solid-svg-icons";
-import { FontAwesomeIcon } from "@fortawesome/vue-fontawesome";
-
-import MarkdownDefault from "../Sections/MarkdownDefault.vue";
-import MarkdownVega from "../Sections/MarkdownVega.vue";
-import MarkdownVitessce from "../Sections/MarkdownVitessce.vue";
 import CellButtonAdd from "./CellButtonAdd.vue";
-import CellCode from "./CellCode.vue";
 import CellWrapper from "./CellWrapper.vue";
 
 interface CellType {
@@ -73,16 +32,6 @@ const props = defineProps<{
 const emit = defineEmits(["update"]);
 
 const cells = ref<Array<CellType>>(parseMarkdown(props.markdownText));
-
-function getMode(cellName: string) {
-    switch (cellName) {
-        case "galaxy":
-            return "python";
-        case "markdown":
-            return "markdown";
-    }
-    return "json";
-}
 
 function onAdd(cellIndex: number, cellType: string) {
     console.log([cellIndex, cellType]);
@@ -111,7 +60,7 @@ function onUpdate() {
 }
 </script>
 
-<style scoped>
+<style>
 .cell-guide {
     width: 3rem;
 }
